requirements.txt
selenium==4.15.2
webdriver-manager==4.0.1
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
python-dotenv==1.0.0

